WIP: Automatic registration of used types in fields/methods
With this change Tria will automatically detect if a type, which is
used in a field or in a method (Return type, argument type), is already
declared using Q_DECLARE_METATYPE(). If it is not, it will do so in the
output. The type will be registered regardless using qRegisterMetaType.
POD- and Qt-types, which are registered by default, will be ignored.

It's not detected yet if the type itself can be registered though.