Added NURIA_READ/NURIA_WRITE and NURIA_REQUIRE for fields.
This (huge) patch adds NURIA_READ/_WRITE macros which can be used to
annotate methods to act as getters/setters for fields. On top of this,
it adds NURIA_REQUIRE, a mechanism which can be used to guard
write-calls against illegal values when writing them through MetaField.
NURIA_REQUIRE expects a condition as argument, which will be embedded
in the output and thus has next to zero overhead.