Fix: Wrong values for elements for enums in JSON output
With the recent introduction of the JSON generator, a 'values' variable
was added to EnumDef. "Definitions" only sorted 'keys', so keys were
given the wrong values. The C++/Nuria generator was not affected as
it only uses the keys and lets the compiler find the values when the
output file is compiled.